refactor: remove all references to ResourceGUID
