+++ Poly/src/Poly/Rendering/RenderGraph/Compiler/RGCContext.h
@@ -4,7 +4,7 @@
 
 #include "Poly/Rendering/RenderGraph/RenderGraph.h"
 #include "Poly/Rendering/RenderGraph/Compiler/CompiledGraph.h"
-#include "Poly/Rendering/RenderGraph/ResourceGUID.h"
+#include "Poly/Rendering/RenderGraph/PassResID.h"
 
 #include <unordered_map>
 #include <vector>
@@ -27,7 +27,7 @@ namespace Poly
 
 		// Written by RGCSynchroniser — final layout/access/stage of every physical
 		// resource after all normal sync passes have been inserted.
-		// Key: canonical ResourceGUID (ResourceCache::GetCanonicalGUID).
+		// Key: canonical PassResID (ResourceCache::GetCanonicalGUID).
 		std::unordered_map<PassResID, ResourceState> PostSyncResourceStates;
 
 		// Written by RGCDebugTextureInjector — GUIDs of texture outputs made
