+++ Poly/src/Poly/Rendering/RenderGraph/RenderData.cpp
@@ -15,13 +15,13 @@ namespace Poly
 
 	Resource* RenderData::GetResourceNonConst(const std::string& resourceName) const
 	{
-		// ResourceGUID is created within the function to avoid the pass to get resources outside of pass scope
+		// PassResID is created within the function to avoid the pass to get resources outside of pass scope
 		return m_pResourceCache->GetResource({ m_RenderPassName, resourceName });
 	}
 
 	const Resource* RenderData::operator[] (const std::string& resourceName) const
 	{
-		// ResourceGUID is created within the function to avoid the pass to get resources outside of pass scope
+		// PassResID is created within the function to avoid the pass to get resources outside of pass scope
 		return m_pResourceCache->GetResource({ m_RenderPassName, resourceName });
 	}
 }
